pyproject.toml
[project]
name = "mcp-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.3.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.1",
]
